# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backtest Platform Research API",
    "version": "0.6.0"
  },
  "paths": {
...
    "/gate/spec": {
      "get": {
        "tags": [
          "gate"
        ],
        "summary": "Gate Spec",
        "description": "Return a gate's criteria (key / op / threshold / kind / label).\n\n``strategy`` \u2192 that strategy's declared gate; omitted \u2192 four-layer DEFAULT_GATE.",
        "operationId": "gate_spec_gate_spec_get",
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GateSpecData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gate/evaluate": {
      "post": {
        "tags": [
          "gate"
        ],
        "summary": "Gate Evaluate",
        "description": "Judge a metrics dict against a gate.\n\n``req.strategy`` \u2192 that strategy's declared gate; omitted \u2192 four-layer default.",
        "operationId": "gate_evaluate_gate_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GateEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GateEvalData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/research/workflows/{strategy}": {
      "get": {
        "tags": [
          "research-workflows"
        ],
        "summary": "List Strategy Workflows",
        "description": "List which workflow configs this strategy declares.",
        "operationId": "list_strategy_workflows_research_workflows__strategy__get",
        "parameters": [
          {
            "name": "strategy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/workflows/{workflow}": {
      "post": {
        "tags": [
          "research-workflows"
        ],
        "summary": "Submit Workflow",
        "description": "Enqueue a research workflow as a background job; returns {job_id, status}.",
        "operationId": "submit_workflow_research_workflows__workflow__post",
        "parameters": [
          {
            "name": "workflow",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_WorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "description": "Structured error body shared by every failure response."
      },
...
      "Envelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope",
        "description": "The single response shape shared by every endpoint.\n\nGeneric over the ``data`` payload: ``response_model=Envelope[RunRow]`` types\nthe OpenAPI ``data`` schema (and the frontend's generated types) per endpoint.\nUnparametrised ``Envelope`` \u2261 ``Envelope[Any]`` \u2014 the ``ok``/``fail``/``pending``\nhelpers and any endpoint without a specific shape keep working untyped, with\nno field loss."
      },
...
      "Envelope_GateEvalData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GateEvalData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[GateEvalData]"
      },
      "Envelope_GateSpecData_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GateSpecData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "Envelope[GateSpecData]"
      },
...
      "GateCriterion": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "op": {
            "type": "string",
            "title": "Op"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "key",
          "op",
          "threshold"
        ],
        "title": "GateCriterion",
        "description": "One ADR-016/019 gate criterion (``GET /gate/spec``)."
      },
      "GateEvalData": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "GateEvalData",
        "description": "``POST /gate/evaluate`` verdict; per-criterion detail via extra='allow'."
      },
      "GateEvaluateRequest": {
        "properties": {
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics",
            "description": "run metrics, e.g. {cagr, sharpe, struct1_pct, ...}"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy",
            "description": "registered strategy name \u2192 judge by ITS declared gate; omit for the four-layer DEFAULT_GATE (\u5be9\u67e5\u7f3a\u9677 #8)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metrics"
        ],
        "title": "GateEvaluateRequest",
        "description": "Judge an arbitrary metrics dict against the\u5be9\u5224\u5ead (ADR-016 + ADR-019)."
      },
      "GateSpecData": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/GateCriterion"
            },
            "type": "array",
            "title": "Criteria"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "GateSpecData"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "_WorkflowRequest": {
        "properties": {
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Overrides",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "strategy"
        ],
        "title": "_WorkflowRequest"
      }
    }
  }
}